+++ qlibs/math/vec.py
@@ -1,14 +1,83 @@
+"""
+## Overview
+
+Vectors are collections of numbers. They can be added, substracted, multiplied and divided, in addition to several useful methods. Because of that they are great for representing locations and directions.
+
+## Creating vectors
+
+- `Vec(iterable)` - creates vector from iterable
+- `Vec(*args)` - creates vector as if args were an iterable, so that `Vec([0, 1])` and `Vec(0, 1)` are equal.
+- `Vec2(x, y)` - specialized vector class, which is faster(direct access to **x** and **y** fields) and is always two-dimensional.
+
+## Indexing
+
+First 4 components of vector can be accessed when they are present like this:
+```python
+from qlibs.math import Vec
+vec = Vec(0, 1, 2, 3)
+print(vec.x, vec.y, vec.z, vec.w)
+#Also possible to index like list
+assert vec.x == vec[0]
+assert vec.y == vec[1]
+assert vec.z == vec[2]
+assert vec.w == vec[3]
+```
+
+## Operators
+```python
+from qlibs.math import Vec
+
+v0 = Vec(10, 5)
+v1 = Vec(0, 20)
+
+#Addition and substraction work elementwise
+print(v0+v1) #Vec(10, 25)
+print(v0-v1) #Vec(10, -15)
+
+#You can multiply or divide vectors by vectors, only by numbers
+#Resulting vector will have elements of first array with operations applied
+print(v0*2)  #Vec(20, 10)
+print(v0/2)  #Vec(5.0, 2.5)
+print(v0//2) #Vec(5, 2)
+
+#Unary minus is also supported
+print(-v0)   #Vec(-10, -5)
+
+#As well as abs()
+print(abs(-v0)) # Vec(10, 5)
+
+#It is also possible to iterate for each element in vector, like list
+for e in v0:
+    print(e)
+
+#Length of the vector (Note: to get number of dimensions len(vec) is required)
+print(v0.len())    #equal to math.hypot(v0.x, v0.y), but also works for more dimensions
+print(v0.len_sq()) #v0.len()**2
+
+#Calculates dot product of vectors
+print(v0.dot(v1))
+
+#Normalize divides vector by it's length, so it will have length of 1
+v0.normalize()
+assert v0.len() == 1
+```
+"""
+
 import math
 from numbers import Number
 from array import array
 
 import warnings
 #from ..net.qpacket import conv_lookup, convert, make_qlibs_obj_id
+__all__ = ["VecBase", "Vec", "Vec2"]
 
 VERTNAMES = {"x": 0, "y": 1, "z": 2, "w": 3}
 NUMERICAL = Number
 
 class VecBase:
+    """
+    Base class for vectors
+    """
     __slots__ = tuple()
     def __eq__(self, other):
         try:
@@ -32,7 +101,10 @@ class VecBase:
         assert len(self) == len(other), "wrong dimensions"
         return self.__class__(*[function(x, y) for x, y in zip(self, other)])
 
-    def dot(self, other):
+    def dot(self, other: "VecBase"):
+        """
+        Vector dot product
+        """
         return sum((v1 * v2 for v1, v2 in zip(self, other)))
 
     def cross(self, other): #Probably need to move this to Vec3
@@ -98,13 +170,22 @@ class VecBase:
         return self.__class__(*map(abs, self))
 
     def normalized(self):
+        """
+        Returns a normalized copy of vector.
+        """
         ln = self.len()
         return self.__class__(*map(lambda x: x / ln, self))
 
     def len_sqr(self):
+        """
+        Returns vector length, squared.
+        """
         return math.fsum(map(lambda x: x ** 2, self))
 
     def len(self):
+        """
+        Returns vector length.
+        """
         return math.sqrt(self.len_sqr())
 
     def to_tuple(self):
@@ -114,6 +195,9 @@ class VecBase:
         return array("f", self).tobytes()
 
     def normalize(self):
+        """
+        Normalizes vector in-place.
+        """
         ln = self.len()
         for i, e in enumerate(self):
             self[i] = e / ln
@@ -126,6 +210,9 @@ class VecBase:
 
 
 class Vec(VecBase):  # Making those immutable is a good idea, right? (No)
+    """
+    Any-dimensional vector. Inherits **VecBase**.
+    """
     __slots__ = ("_v",)
 
     def __init__(self, *args):
@@ -174,6 +261,9 @@ class Vec(VecBase):  # Making those immutable is a good idea, right? (No)
 
 
 class Vec2(VecBase):
+    """
+    2-dimensional vector. Faster than **Vec**. Inherits **VecBase**.
+    """
     __slots__ = ("x", "y")
     def __init__(self, x, y):
         self.x = x
@@ -220,6 +310,9 @@ class Vec2(VecBase):
         self.x, self.y = math.cos(value)*l, math.sin(value)*l
     
     def rotate(self, angle) -> "Vec2":
+        """
+        Rotate vector by *angle* degrees
+        """
         c = self.angle + angle
         l = self.len()
         return Vec2(math.cos(c)*l, math.sin(c)*l)